Updated install.sh

+++ install.sh
@@ -2,9 +2,13 @@
 
 echo "INSTALLING BASHTOOLS"
 echo
-
-echo -n "Update system first (apt full-upgrade)? y/N "
-read res
+echo "Note: To upgrade to a specific OS release:"
+echo "      - sudo nano /etc/apt/sources.list"
+echo "      - Modify version name (replace (e.g. strech with bullseye)"
+echo "      - Save and close with STRG+X"
+echo
+echo -n "Should we perform a system update now (apt full-upgrade)? y/N "
+read -n1 res
 if [ $res == "y" ]
 then
     echo "Performing system update..."
@@ -12,6 +16,9 @@ then
     sudo apt update
     sudo apt -y full-upgrade
     sudo apt -y autoremove
+    sudo apt -y autoclean
+    echo
+    lsb_release -a
 fi
 
 echo
@@ -20,7 +27,11 @@ echo
 
 sudo apt install freerdp2-x11 moreutils
 
-cat > .env
+if [ ! -f .env ]
+then
+    echo
+    echo "Creating empty .env file"
+    cat > .env
 host=
 user=
 password=
@@ -28,6 +39,7 @@ resolution=
 bitdepth=
 logfile=
 <EOF>
+fi
 
 echo
 echo "Done!"
